1-create_file.c

+++ 0x15-file_io/1-create_file.c
@@ -11,7 +11,7 @@
 
 int create_file(const char *filename, char *text_content)
 {
-	unsigned int fd;
+	long int fd;
 
 	if (filename == NULL)
 		return (-1);
